requirements.txt
apache-airflow
requests
boto3
